src/main/java/ru/ssau/tk/labochkimoi/smirnovkondorova/operations/TabulatedFunctionOperationService.java
package ru.ssau.tk.labochkimoi.smirnovkondorova.operations;
import ru.ssau.tk.labochkimoi.smirnovkondorova.functions.TabulatedFunction;
import ru.ssau.tk.labochkimoi.smirnovkondorova.functions.Point;
public class TabulatedFunctionOperationService {
    public static Point[] asPoints(TabulatedFunction tabulatedFunction) {
        Point[] points = new Point[tabulatedFunction.getCount()];
        int i = 0;
        for (Point point : tabulatedFunction) {
            points[i++] = point;
        }
        return points;
    }
}
